Fixed installer + update function

+++ script/openssl.sh
@@ -51,37 +51,51 @@ make install >>"${make_log}" 2>>"${make_err_log}"
 
 cd ~/sources/
 rm -R openssl-${OPENSSL_VERSION}
+cd ~/
 }
 
 update_openssl() {
 
-mkdir -p ~/sources/
+source configs/versions.cfg
 
-apt-get install -y libtool zlib1g-dev libpcre3-dev libssl-dev libxslt1-dev libxml2-dev libgd2-xpm-dev libgeoip-dev libgoogle-perftools-dev libperl-dev >>"${main_log}" 2>>"${err_log}"
+LOCAL_OPENSSL_VERSION_STRING=$(openssl version)
+LOCAL_OPENSSL_VERSION=$(echo $LOCAL_OPENSSL_VERSION_STRING | cut -c9-14)
 
-cd ~/sources
-wget -c4 --no-check-certificate https://www.openssl.org/source/openssl-${OPENSSL_VERSION}.tar.gz --tries=3 >>"${main_log}" 2>>"${err_log}"
-	ERROR=$?
-	if [[ "$ERROR" != '0' ]]; then
-      echo "Error: openssl-${OPENSSL_VERSION}.tar.gz download failed."
-      exit
-    fi
+if [[ ${LOCAL_OPENSSL_VERSION} != ${OPENSSL_VERSION} ]]; then
 
-tar -xzf openssl-${OPENSSL_VERSION}.tar.gz >>"${main_log}" 2>>"${err_log}"
-	ERROR=$?
-	if [[ "$ERROR" != '0' ]]; then
-      echo "Error: openssl-${OPENSSL_VERSION}.tar.gz is corrupted."
-      exit
-    fi
-rm openssl-${OPENSSL_VERSION}.tar.gz	
+	mkdir -p ~/sources/
+
+	cd ~/sources
+	wget -c4 --no-check-certificate https://www.openssl.org/source/openssl-${OPENSSL_VERSION}.tar.gz --tries=3 >>"${main_log}" 2>>"${err_log}"
+		ERROR=$?
+		if [[ "$ERROR" != '0' ]]; then
+		  echo "Error: openssl-${OPENSSL_VERSION}.tar.gz download failed."
+		  exit
+		fi
+
+	tar -xzf openssl-${OPENSSL_VERSION}.tar.gz >>"${main_log}" 2>>"${err_log}"
+		ERROR=$?
+		if [[ "$ERROR" != '0' ]]; then
+		  echo "Error: openssl-${OPENSSL_VERSION}.tar.gz is corrupted."
+		  exit
+		fi
+	rm openssl-${OPENSSL_VERSION}.tar.gz	
+		
+	cd openssl-${OPENSSL_VERSION}
+
+	./config --prefix=/usr --openssldir=/etc/ssl --libdir=lib shared zlib-dynamic >>"${make_log}" 2>>"${make_err_log}"
+			 
+	make -j $(nproc) >>"${make_log}" 2>>"${make_err_log}"
+	make install >>"${make_log}" 2>>"${make_err_log}"
+
+	cd ~/sources/
+	rm -R openssl-${OPENSSL_VERSION}
+	cd ~/
 	
-cd openssl-${OPENSSL_VERSION}
+else
 
-./config --prefix=/usr --openssldir=/etc/ssl --libdir=lib shared zlib-dynamic >>"${make_log}" 2>>"${make_err_log}"
-		 
-make -j $(nproc) >>"${make_log}" 2>>"${make_err_log}"
-make install >>"${make_log}" 2>>"${make_err_log}"
-
-cd ~/sources/
-rm -R openssl-${OPENSSL_VERSION}
+	HEIGHT=10
+	WIDTH=70
+	dialog --backtitle "Welcome to the NeXt Server installation!" --infobox "No Openssl Update needed! Local Openssl Version: ${LOCAL_OPENSSL_VERSION}. Version to be installed: ${OPENSSL_VERSION}" $HEIGHT $WIDTH
+fi	
 }
